improve error message for invalid data 🥳
The attribute `#[serde(untagged)]` on the `attending` field is the
correct one for how the data was originally modeled, but it comes with
a downside. If there is any deserialization error within the `attending`
structure, the resulting error message won't be very useful. For
example, if there is an incorrect `AttendingValue` like "dINNER", the
resulting error will be:

> data did not match any variant of untagged enum AttendingField

The only way to solve this is to add a custom Deserialize implementation
that commits to one of the enum variants earlier. In this case, we can
commit based on whether the value is a list or a map.

For the above example of "dINNER", the new error message is:

> unknown variant `dINNER`, expected one of `afternoon`, `dinner`, `hike`

This also improves the error message if `attending` is neither a list
nor a map, for example `attending = 12`:

> invalid type: integer `12`, expected a list of AttendingValue or a map
> of strings (names of sub-attendees) to list of AttendingValue